refactor: extract BlueScreen and BootSequence into own components
BlueScreen.vue   — presentational; renders the BSOD blue screen with
                   a countdown prop.
BootSequence.vue — self-contained sequence (BSOD → BIOS → CD boot →
                   Winders 98 start); runs entirely on onMounted,
                   emits 'done' when finished. Uses <BlueScreen>
                   internally for the BSOD phase.
Win98Window.vue  — now just mounts <BootSequence v-if="showBoot">
                   and calls emit('close') on its 'done' event;
                   all content arrays, phase state and timing
                   logic removed.
